test(e2e): add canister ids as env vars for e2e tests
