+++ playwright.config.ts
@@ -1,4 +1,15 @@
 import { defineConfig, devices } from '@playwright/test';
+import dotenv, { type DotenvPopulateInput } from 'dotenv';
+import { join } from 'node:path';
+import { readCanisterIdsFromJSONFile } from './shared.utils';
+
+dotenv.populate(
+	process.env as DotenvPopulateInput,
+	readCanisterIdsFromJSONFile({
+		filePath: join(process.cwd(), 'canister_e2e_ids.json'),
+		prefix: 'E2E_'
+	})
+);
 
 const DEV = (process.env.NODE_ENV ?? 'production') === 'development';
