+++ tsconfig.eslint.json
@@ -9,6 +9,7 @@
 		"./src/**/*.js",
 		"./src/**/*.ts",
 		"./src/**/*.svelte",
+		"./shared.utils.ts",
 		"./vitest.setup.ts",
 		"./vitest.config.ts",
 		"./tailwind.config.js",
