+++ vite.utils.ts
@@ -1,47 +1,7 @@
 import { existsSync, readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import { fileURLToPath } from 'node:url';
-
-const readIds = ({
-	filePath,
-	prefix
-}: {
-	filePath: string;
-	prefix?: string;
-}): Record<string, string> => {
-	try {
-		type Details = {
-			ic?: string;
-			staging?: string;
-			local?: string;
-		};
-
-		const config: Record<string, Details> = JSON.parse(readFileSync(filePath, 'utf8'));
-
-		return Object.entries(config).reduce((acc, current: [string, Details]) => {
-			const [canisterName, canisterDetails] = current;
-
-			const ids = Object.entries(canisterDetails).reduce(
-				(acc, [network, id]) => ({
-					...acc,
-					[`${prefix ?? ''}${network.toUpperCase()}_${canisterName
-						.replaceAll('-', '_')
-						.replaceAll("'", '')
-						.toUpperCase()}_CANISTER_ID`]: id
-				}),
-				{}
-			);
-
-			return {
-				...acc,
-				...ids
-			};
-		}, {});
-	} catch (e) {
-		console.warn(`Could not get canister ID from ${filePath}: ${e}`);
-		return {};
-	}
-};
+import { readCanisterIdsFromJSONFile } from './shared.utils';
 
 /**
  * Read all the locally deployed canister IDs. For example Oisy backend, ckBTC|ETH, ICP etc.
@@ -50,7 +10,7 @@ const readIds = ({
 const readLocalCanisterIds = ({ prefix }: { prefix?: string }): Record<string, string> => {
 	const dfxCanisterIdsJsonFile = join(process.cwd(), '.dfx', 'local', 'canister_ids.json');
 	const e2eCanisterIdsJsonFile = join(process.cwd(), 'canister_e2e_ids.json');
-	return readIds({
+	return readCanisterIdsFromJSONFile({
 		filePath: existsSync(dfxCanisterIdsJsonFile) ? dfxCanisterIdsJsonFile : e2eCanisterIdsJsonFile,
 		prefix
 	});
@@ -62,7 +22,7 @@ const readLocalCanisterIds = ({ prefix }: { prefix?: string }): Record<string, s
  */
 const readOisyCanisterIds = ({ prefix }: { prefix?: string }): Record<string, string> => {
 	const canisterIdsJsonFile = join(process.cwd(), 'canister_ids.json');
-	return readIds({ filePath: canisterIdsJsonFile, prefix });
+	return readCanisterIdsFromJSONFile({ filePath: canisterIdsJsonFile, prefix });
 };
 
 /**
